requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anthropic==0.2.10
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
dataclasses-json==0.5.7
distro==1.8.0
exceptiongroup==1.1.2
frozenlist==1.3.3
google-search-results==2.4.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
langchain==0.0.196
langchainplus-sdk==0.0.8
langsmith==0.0.5
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.9
python-decouple==3.8
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.16
tenacity==8.2.2
tokenizers==0.13.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
yarl==1.9.2
